LPS-54223 migrate monitoring.jmx classes to module
